client/requirements.txt
picamera==1.12
Pillow==4.0.0
RPi.GPIO==0.6.3
python-nmap==0.6.1
